requirements.txt
pillow
numpy
progress